+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Added
+- Connect software wallet flow (export extended public key via QR)
+
+## [0.1.0] - 2025-02-22
+
+### Added
 - Scanning QR code from compatible Ethereum wallets
 - Signing transaction with Keycard
-- Scan back result QR code into the compatible Ethereum wallet 
+- Scan back result QR code into the compatible Ethereum wallet
+
+[Unreleased]: https://github.com/mmlado/GapSign/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/mmlado/GapSign/compare/294c1212cfd8d1738b5eb90bbb33aa02adee139c...v0.1.0